src/test/java/com/example/healthcare/controller/UserControllerTest.java
package com.example.healthcare.controller;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import com.example.healthcare.model.User;
import com.example.healthcare.services.UserService;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

@WebMvcTest(UserController.class)
public class UserControllerTest {

  @Autowired
  private MockMvc mockMvc;

  @MockBean
  private UserService userService;

  @Test
  public void testRegisterUser_Success() throws Exception {
    User user = new User();
    user.setUsername("testuser");
    user.setPassword("password");

    when(userService.registerUser(any(User.class))).thenReturn(user);

    mockMvc.perform(post("/api/users/register")
            .contentType(MediaType.APPLICATION_JSON)
            .content(new ObjectMapper().writeValueAsString(user)))
        .andExpect(status().isOk())
        .andExpect(content().string("User registered successfully"));
  }

  @Test
  public void testRegisterUser_UsernameExists() throws Exception {
    User user = new User();
    user.setUsername("existinguser");
    user.setPassword("password");

    when(userService.registerUser(any(User.class))).thenThrow(
        new RuntimeException("Username already exists"));

    mockMvc.perform(post("/api/users/register")
            .contentType(MediaType.APPLICATION_JSON)
            .content(new ObjectMapper().writeValueAsString(user)))
        .andExpect(status().isBadRequest());
  }
}
